Basic Database and RestAPI done. Navigation function still developing.

+++ src/main/java/edu/nus/campus/model/Stop.java
@@ -3,8 +3,8 @@ package edu.nus.campus.model;
 public class Stop {
     private int id;
     private String name;
-    private int pos_x;
-    private int pos_y;
+    private float pos_x;
+    private float pos_y;
 
 
     // Default getters and setters;
@@ -24,19 +24,19 @@ public class Stop {
         this.name = name;
     }
 
-    public int getPos_x() {
+    public float getPos_x() {
         return pos_x;
     }
 
-    public void setPos_x(int pos_x) {
+    public void setPos_x(float pos_x) {
         this.pos_x = pos_x;
     }
 
-    public int getPos_y() {
+    public float getPos_y() {
         return pos_y;
     }
 
-    public void setPos_y(int pos_y) {
+    public void setPos_y(float pos_y) {
         this.pos_y = pos_y;
     }
 }
